Register page implemented

+++ studybud/base/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
@@ -31,12 +32,15 @@ def home(request):
 
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
 
         try :
-            user = User.objects.get(username=username)
+            user = User.objects.get(username=username).lower()
             print(user)
         except User.DoesNotExist:
             messages.error(request, "User does not exist")
@@ -48,8 +52,30 @@ def loginPage(request):
         else:
             messages.error(request, "Authentication failed")
             
-    context = {}
+    context = {'page': page}
     return render(request,'base/login_reg.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    page = 'register'
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occurred")
+    else:
+        form = UserCreationForm()
+    return render(request, 'base/login_reg.html', {'form': form, 'page': page})
+
 def room(request,pk):
     room = Room.objects.get(id=pk)
 
@@ -57,6 +83,7 @@ def room(request,pk):
     context = {'room':room}
     return render(request,'base/room.html',context)
 
+@login_required(login_url="/login")
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -68,9 +95,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request,'base/room_form.html', context)
 
+@login_required(login_url="/login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=int(pk))
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -82,8 +113,12 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request,'base/room_form.html', context)
 
+@login_required(login_url="/login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
